renaming variables and adjusting formula for ma_eaedc and ma_tafdc
