+++ policyengine_us/tests/policy/baseline/gov/states/ma/dta/tcap/eaedc/income/ma_eaedc_earned_income_after_disregard_person.yaml
@@ -50,7 +50,7 @@
     ma_tcap_gross_earned_income: 9_600
     state_code: MA
   output:
-    ma_eaedc_earned_income_after_disregard_person: 6087.6
+    ma_eaedc_earned_income_after_disregard_person: 6_087.6
     # 800 - 200 = 600
     # Percentage disregard = (600 - 30)*0.33*4 = 752.4
     # Flat disregard = 30*12 = 360
