+++ policyengine_us/tests/policy/baseline/gov/states/ma/dta/tcap/eaedc/income/ma_eaedc_standard_assistance.yaml
@@ -68,4 +68,4 @@
     spm_unit_size: 2 
     state_code: MA
   output:
-    ma_eaedc_standard_assistance: 6264 # 522*12
+    ma_eaedc_standard_assistance: 6_264 # 522*12
