+++ policyengine_us/tests/policy/baseline/gov/states/ma/dta/tcap/eaedc/ma_eaedc.yaml
@@ -159,14 +159,14 @@
         state_code: MA
         #is_in_public_housing: true
   output:
-    ma_tcap_gross_earned_income: [9600, 0] # 800 per month
-    ma_tafdc_countable_earned_income: [1800, 0]
-    ma_tafdc_applicable_income_grant_amount: 1800
+    ma_tcap_gross_earned_income: [9_600, 0] # 800 per month
+    ma_tafdc_countable_earned_income: [1_800, 0]
+    ma_tafdc_applicable_income_grant_amount: 1_800
     ma_tafdc_eligible: true
     ma_tafdc_payment_standard: 8_256 #688*12
     ma_tafdc_eligible_dependent: [False, True]
     ma_tafdc_dependent_care_deduction: 0
-    ma_tafdc: 6956 # 6959/12 = 580
+    ma_tafdc: 6_956 # 6959/12 = 580
     ma_eaedc_eligible: false
     ma_eaedc_immigration_status_eligible: true
     ma_eaedc_financial_eligible: false
