+++ policyengine_us/tests/policy/baseline/gov/states/ma/dta/tcap/eaedc/ma_eaedc_if_tafdc_claimed.yaml
@@ -5,7 +5,7 @@
     ma_eaedc_net_income: 80
     ma_eaedc_eligible: true
   output:
-    ma_eaedc_if_tafdc_ineligible: 20
+    ma_eaedc_if_claimed: 20
 
 - name: Standard assistance lower than income.  
   period: 2023
@@ -14,4 +14,4 @@
     ma_eaedc_net_income: 100
     ma_eaedc_eligible: true
   output:
-    ma_eaedc_if_tafdc_ineligible: 0
+    ma_eaedc_if_claimed: 0
