+++ policyengine_us/tests/policy/baseline/gov/states/ma/dta/tcap/tafdc/ma_tafdc.yaml
@@ -3,10 +3,10 @@
   input:
     ma_tafdc_payment_standard: 2_100
     ma_tafdc_clothing_allowance: 400
+    ma_tafdc_exceeds_eaedc: true
     ma_tafdc_eligible: true
     ma_tafdc_applicable_income_grant_amount: 2_200
   output:
-  # This case will be neglected in integration tests due to income eligibility.
     ma_tafdc: 400
 
 - name: Countable income under payment standard
@@ -14,6 +14,7 @@
   input:
     ma_tafdc_payment_standard: 2_100
     ma_tafdc_clothing_allowance: 400
+    ma_tafdc_exceeds_eaedc: true
     ma_tafdc_eligible: true
     ma_tafdc_applicable_income_grant_amount: 1_000
   output:
@@ -24,8 +25,9 @@
   input:
     ma_tafdc_payment_standard: 2_100
     ma_tafdc_clothing_allowance: 400
+    ma_tafdc_exceeds_eaedc: false
     ma_tafdc_eligible: false
-    ma_tafdc_applicable_income_grant_amount: 2_500
+    ma_tafdc_applicable_income_grant_amount: 2_400
   output:
     ma_tafdc: 0
 
@@ -54,14 +56,14 @@
         state_code: MA
         #is_in_public_housing: true
   output:
-    ma_tcap_gross_earned_income: [9600, 0] # 800 per month
-    ma_tafdc_countable_earned_income: [1800, 0]
-    ma_tafdc_applicable_income_grant_amount: 1800
-    ma_tafdc_eligible: true
+    ma_tcap_gross_earned_income: [9_600, 0] # 800 per month
+    ma_tafdc_countable_earned_income: [1_800, 0]
+    ma_tafdc_applicable_income_grant_amount: 1_800
+    ma_tafdc_exceeds_eaedc: true
     ma_tafdc_payment_standard: 8_256 #688*12
     ma_tafdc_eligible_dependent: [False, True]
     ma_tafdc_dependent_care_deduction: 0
-    ma_tafdc: 6956 # 6959/12 = 580
+    ma_tafdc: 6_956 # 6959/12 = 580
     ma_eaedc_eligible: false
     ma_eaedc_immigration_status_eligible: true
     ma_eaedc_financial_eligible: false
