+++ policyengine_us/tests/policy/baseline/gov/states/ma/dta/tcap/tafdc/ma_tafdc_if_claimed.yaml
@@ -1,0 +1,23 @@
+- name: Payment standard higher than income. 
+  period: 2023
+  input:
+    ma_tafdc_payment_standard: 100
+    ma_tafdc_applicable_income_grant_amount: 80
+    ma_tafdc_clothing_allowance: 0 
+    ma_tafdc_infant_benefit: 0
+    ma_tafdc_eligible: true
+    state_code: MA
+  output:
+    ma_tafdc_if_claimed: 20
+
+- name: Payment standard lower than income
+  period: 2023
+  input:
+    ma_tafdc_payment_standard: 100
+    ma_tafdc_applicable_income_grant_amount: 180
+    ma_tafdc_clothing_allowance: 0 
+    ma_tafdc_infant_benefit: 0
+    ma_tafdc_eligible: true
+    state_code: MA
+  output:
+    ma_tafdc_if_claimed: 0
