--- policyengine_us/tests/policy/baseline/gov/states/ma/dta/tcap/tafdc/ma_tafdc_larger_than_eaedc.yaml
@@ -1,17 +1,0 @@
-- name: TAFDC larger than EAEDC.
-  period: 2023
-  input: 
-    ma_tafdc_if_eaedc_ineligible: 100
-    ma_eaedc_if_tafdc_ineligible: 80
-    state_code: MA
-  output:
-    ma_tafdc_larger_than_eaedc: true
-
-- name: TAFDC lower than EAEDC.
-  period: 2023
-  input: 
-    ma_tafdc_if_eaedc_ineligible: 100
-    ma_eaedc_if_tafdc_ineligible: 180
-    state_code: MA
-  output:
-    ma_tafdc_larger_than_eaedc: false
