+++ policyengine_us/tests/policy/baseline/gov/states/ma/dta/tcap/tafdc/ma_tafdc_potential_main_benefit.yaml
@@ -3,21 +3,17 @@
   input:
     ma_tafdc_payment_standard: 100
     ma_tafdc_applicable_income_grant_amount: 80
-    ma_tafdc_clothing_allowance: 0 
-    ma_tafdc_infant_benefit: 0
     ma_tafdc_eligible: true
     state_code: MA
   output:
-    ma_tafdc_if_eaedc_ineligible: 20
+    ma_tafdc_potential_main_benefit: 20
 
 - name: Payment standard lower than income
   period: 2023
   input:
     ma_tafdc_payment_standard: 100
     ma_tafdc_applicable_income_grant_amount: 180
-    ma_tafdc_clothing_allowance: 0 
-    ma_tafdc_infant_benefit: 0
     ma_tafdc_eligible: true
     state_code: MA
   output:
-    ma_tafdc_if_eaedc_ineligible: 0
+    ma_tafdc_potential_main_benefit: 0
