+++ policyengine_us/variables/gov/states/ma/dta/tcap/eaedc/ma_eaedc.py
@@ -10,11 +10,7 @@ class ma_eaedc(Variable):
     defined_for = "ma_eaedc_eligible"
     reference = "https://www.law.cornell.edu/regulations/massachusetts/department-106-CMR/title-106-CMR-701.000"
 
-    # Program value can not be less than 0
-    # due to the eligibility requirements
     def formula(spm_unit, period, parameters):
-        tafdc_exceeds_eaedc = spm_unit("ma_tafdc_larger_than_eaedc", period)
-        standard_assistance = spm_unit("ma_eaedc_standard_assistance", period)
-        net_income = spm_unit("ma_eaedc_net_income", period)
-        eaedc_benefit_amount = max_(standard_assistance - net_income, 0)
+        tafdc_exceeds_eaedc = spm_unit("ma_tafdc_exceeds_eaedc", period)
+        eaedc_benefit_amount = spm_unit("ma_eaedc_if_claimed", period)
         return ~tafdc_exceeds_eaedc * eaedc_benefit_amount
