+++ policyengine_us/variables/gov/states/ma/dta/tcap/ma_tafdc_exceeds_eaedc.py
@@ -1,7 +1,7 @@
 from policyengine_us.model_api import *
 
 
-class ma_tafdc_larger_than_eaedc(Variable):
+class ma_tafdc_exceeds_eaedc(Variable):
     value_type = bool
     entity = SPMUnit
     label = "Whether the TAFDC value exceeds the EAEDC value"
@@ -10,6 +10,6 @@ class ma_tafdc_larger_than_eaedc(Variable):
     defined_for = StateCode.MA
 
     def formula(spm_unit, period, parameters):
-        tafdc_value = spm_unit("ma_tafdc_if_eaedc_ineligible", period)
-        eaedc_value = spm_unit("ma_eaedc_if_tafdc_ineligible", period)
+        tafdc_value = spm_unit("ma_tafdc_if_claimed", period)
+        eaedc_value = spm_unit("ma_eaedc_if_claimed", period)
         return tafdc_value > eaedc_value
